pyproject.toml
[project]
name = "trafficlight"
version = "0.1.0"
description = "traffic light detection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.135",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
